+++ lib/simple_jwt_manager.dart
@@ -1,8 +1,8 @@
 export 'src/manager/timing_manager.dart';
 export 'src/network/enum_post_type.dart';
 export 'src/network/util_check_url.dart';
-export 'src/network/util_https_stub.dart'
-    if (dart.library.js_interop) 'src/network/util_https_web.dart';
+export 'src/network/https_service_stub.dart'
+    if (dart.library.js_interop) 'src/network/https_service_web.dart';
 export 'src/server_response/enum_server_response_status.dart';
 export 'src/server_response/enum_server_response_type.dart';
 export 'src/server_response/server_response.dart';