+++ lib/src/network/https_service_stub.dart
@@ -1,0 +1,2 @@
+export 'https_service_for_native.dart';
+export 'https_service.dart';